+++ web/pages/item-requests.tsx
@@ -53,15 +53,15 @@ function genItemRequestsTab({ itemRequests }: { itemRequests: RequestItem[] }) {
 
   return (
     <div>
-        <div class="absolute mt-20 border-t w-full"/>
-        <div class="pt-6 px-8 flex h-full flex-col justify-left">
-            <div class="flex flex-row">
-                <h1 className="text-2xl mb-5 font-bold">Item Requests</h1>
-            </div>
-            <div>
-                <ItemRequestsTable columns={columns} data={data} />
-            </div>
+      <div className="absolute mt-20 border-t w-full" />
+      <div className="pt-6 px-8 flex h-full flex-col justify-left">
+        <div className="flex flex-row">
+          <h1 className="text-2xl mb-5 font-bold">Item Requests</h1>
         </div>
+        <div>
+          <ItemRequestsTable columns={columns} data={data} />
+        </div>
+      </div>
     </div>
   );
 }
